split of electoral_calculus_constituencies_tsv
...constituencies_tsv.rb now purely encapsulates the TSV
...polls_raw.rb does some mapping to SMV ids

Suspect after refactoring and applying fixes for missing data,
that ...polls_raw.rb and ...polls.rb will be mergeable ... they are both
about bridging the gap between EC data and SMV data.
